requirements.txt
Flask
Flask-Cors
requests
streamlit
langchain
langchain_community
chromadb
sentence-transformers
google-generativeai